+++ packages/pangea-multipass/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Handle null fields on issues in JiraME
 - Handle trailing slash in Jira URL
+- GitLabProcessor `get_filter()`
 
 
 ## 0.2.0 - 2025-01-15